cmd/libsnap: make sanity timeout APIs public and move them to locking
The sanity timeout APIs belong logically with the locking primitives
so move them there. In addition, since they are needed in the namespace
module, make them public.
